Removed unnecessary casts (and a few related brackets).
git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2609 c5679a2a-0e0e-0410-bba2-8642b794b8c1

+++ Colossus/core/src/main/java/net/sf/colossus/client/Client.java
@@ -1339,7 +1339,7 @@ public final class Client implements IClient, IOracle, IOptions
         {
             colorName = getColorByMarkerId(attackerMarkerId);
         }
-        BattleChit chit = new BattleChit((int)(5 * Scale.get()), imageName,
+        BattleChit chit = new BattleChit(5 * Scale.get(), imageName,
             inverted, tag, hexLabel, colorName, this);
         battleChits.add(chit);
     }
